renamed multiagent_ring -> lord_of_the_rings
