+++ examples/exp_configs/rl/multiagent/lord_of_the_rings.py
