+++ tests/fast_tests/test_examples.py
@@ -17,8 +17,8 @@ from examples.exp_configs.rl.singleagent.singleagent_bottleneck import flow_para
 from examples.exp_configs.rl.multiagent.adversarial_figure_eight import flow_params as adversarial_figure_eight
 from examples.exp_configs.rl.multiagent.multiagent_figure_eight import flow_params as multiagent_figure_eight
 from examples.exp_configs.rl.multiagent.multiagent_merge import flow_params as multiagent_merge
-from examples.exp_configs.rl.multiagent.multiagent_ring import \
-    flow_params as multiagent_ring
+from examples.exp_configs.rl.multiagent.lord_of_the_rings import \
+    flow_params as lord_of_the_rings
 from examples.exp_configs.rl.multiagent.multiagent_traffic_light_grid import \
     flow_params as multiagent_traffic_light_grid
 from examples.exp_configs.rl.multiagent.multiagent_highway import flow_params as multiagent_highway
@@ -190,17 +190,17 @@ class TestRllibExamples(unittest.TestCase):
         }
         self.run_exp(multiagent_figure_eight, **kwargs)
 
-    def test_multi_ring(self):
-        from examples.exp_configs.rl.multiagent.multiagent_ring import POLICY_GRAPHS as mrpg
-        from examples.exp_configs.rl.multiagent.multiagent_ring import POLICIES_TO_TRAIN as mrpt
-        from examples.exp_configs.rl.multiagent.multiagent_ring import policy_mapping_fn as mrpmf
+    def test_lord_of_the_rings(self):
+        from examples.exp_configs.rl.multiagent.lord_of_the_rings import POLICY_GRAPHS as ltrpg
+        from examples.exp_configs.rl.multiagent.lord_of_the_rings import POLICIES_TO_TRAIN as ltrpt
+        from examples.exp_configs.rl.multiagent.lord_of_the_rings import policy_mapping_fn as ltrpmf
 
         kwargs = {
-            "policy_graphs": mrpg,
-            "policies_to_train": mrpt,
-            "policy_mapping_fn": mrpmf
+            "policy_graphs": ltrpg,
+            "policies_to_train": ltrpt,
+            "policy_mapping_fn": ltrpmf
         }
-        self.run_exp(multiagent_ring, **kwargs)
+        self.run_exp(lord_of_the_rings, **kwargs)
 
     def test_multiagent_merge(self):
         from examples.exp_configs.rl.multiagent.multiagent_merge import POLICY_GRAPHS as mmpg
